api: add fn for creating generator for recursion
When generating recursive structures, users will often use the size of
the generator to bound the recursion depth. Dividing the size by the
number of sub-structures appears to be a common practice. However,
modifying the size of the generator passed to an implementation of
`Arbitrary::arbitrary` would also affect the generation of sister
structures if not restored after the generation of child nodes.

Instead, users can create a new generator instance with the target size.
Historically, this would be the only option since public API for setting
the size of an existing generator was only introduced recently. This
change introduces a convenience function for creating such a generator.

The new function makes use of seed from the original generator rather
than entropy. In addition to bounding recursive structures, this also
allows to re-create those structures based on some initial seed.